Add TestBarrier API helper class
This class implements a simple cyclic barrier for concurrent unit
tests.  Simple means that it has a single await() method, throws no
InterruptedExceptions, and it will be initialized by the testing
framework to wait for all threads of the concurrent unit test;
minimizing the amount of code necessary to use it in a concurrent unit
test.

Signed-off-by: SZEDER Gábor <szeder@fzi.de>