ExceptionWrapper: fix stackTrace lazy initialization
Lazy initialization was not working, because getStackTrace() shadowed
the stackTrace member variable and, therefore, the stack trace was not
stored but was recreated during every getStackTrace() invocation.
(Note: it seems that getStackTrace() is invoked only once for an
exception, but still...)

Too bad we can't unit test it because of the heavy JPF dependencies.

Signed-off-by: SZEDER Gábor <szeder@fzi.de>