ExceptionComparator: compare MultipleFailureExceptions
Even when multiple exceptions are thrown during a concurrent unit
test, we still have to tell whether that is a concurrency or a
sequential bug.  To do that we must compare the list of individual
exception instances stored in MultipleFailureExceptions.  It's not
necessary to compare the MultipleFailureException instances
themselves, because they can't have a message or cause, and there will
always have the same stack trace, because there will be only a single
place in the code where they get thrown from, so they will always be
equal.

Signed-off-by: SZEDER Gábor <szeder@fzi.de>