ExceptionFactory: support reconstruction of MultipleFailureException
Reconstruction of MultipleFailureExceptions is different from
reconstruction other exceptions, because we must also reconstruct the
exceptions stored in the original MultipleFailureException instance,
and pass them as a List to MultipleFailureException's constructor.

This patch adds the createMultipleFailureException() method to
ExceptionFactory to do just that, and implements
MultipleFailureException.reconstruct() to invoke this method.

Signed-off-by: SZEDER Gábor <szeder@fzi.de>