Integration test to check the handling of multiple failures
Currently, if multiple exceptions are thrown during a test, we only
report the first one wrapped in an Exception telling the user that
there were multiple failures, and when the test failed because of a
concurrency bug, then this Exception was wrapped in a ConcurrentError.

Here we add an integration test that shows this behaviour.  Once we
will properly handle MultipleFailureExceptions this integration test
will be amended.

Signed-off-by: SZEDER Gábor <szeder@fzi.de>