Move code collecting information about exceptions to ExceptionInfoCollector
Gaining information about the execption thrown inside JPF is mainly
implemented by the ExceptionWrapper and StackTraceElementInfoWrapper
classes, and ExceptionInfoCollector did almost nothing except to find
the ElementInfo representing the exception to be reconstructed.
However, this turned out to be not as flexible as needed to handle and
reconstruct MultipleFailreExceptions and uncaught exceptions
(especially when they are chained exceptions).

So this patch extends ExceptionInfoCollector to implement the
functionality of the ExceptionWrapper and StackTraceElementInfoWrapper
classes directly on top of ElementInfoWrapper, and removes
ExceptionWrapper and StackTraceElementInfoWrapper since they will not
be used anymore.

Signed-off-by: SZEDER Gábor <szeder@fzi.de>