IntegrationTest: fix message of failure count assertions
The return value of getFailureCount() is not the number of failed
tests, but the number of failures, i.e. the number of exceptions
thrown during the test run.  The two can differ when a test throws
more than one exception (e.g. one by the test method and one by an
after method).

Consequently, the current "number of failed tests" assertion message
is a bit misleading.  Let's change it to "number of failures".

Signed-off-by: SZEDER Gábor <szeder@fzi.de>