Add MultipleFailureExceptionInfo class
MultipleFailureExceptions have a field containing the list of
Throwables which caused the unit test to fail.  We need to know what
these Throwables are in order to reconstruct a
MultipleFailureException outside of JPF.  Therefore, we subclass the
ExceptionInfo class to extend it with an array containing the
ExceptionInfos of those Throwables, to make access to them from
outside of JPF easier.

Signed-off-by: SZEDER Gábor <szeder@fzi.de>