TestWrapper: throw MultipleFailureException in case of multiple exceptions
Now everything is in place to make reconstruction of
MultipleFailureExceptions possible, so these are the final bits to
support reporting of multiple failures in a single concurrent unit
test.

Also update the related integration test to check that both exceptions
causing the failure of the concurrent unit test are reported.

Signed-off-by: SZEDER Gábor <szeder@fzi.de>