ElementInfoWrapper: allow wrapping subclasses
ElementInfoWrapper's constructor currently checks that the class to be
wrapped exactly matches the expected class' type, so no subtypes were
allowed.  However, we need to collect information about an exception
directly from its JPF representation, but we don't know the exact type
of the exception, just that it is a subclass of Throwable.

So this patch loosens that check to allow wrapping subclasses of the
expected class.

Signed-off-by: SZEDER Gábor <szeder@fzi.de>