ExceptionInfo: add reconstruct() method
Currently we have only one ExceptionInfo class, and only one way to
reconstruct it.  However, in the future there will be more
ExceptionInfo classes (at least one to support
MultipleFailureExceptions), and they need to be reconstructed
differently.

To prepare for that this patch extends the ExceptionInfo class with
the reconstruct() method, which will call the appropriate
reconstruction method in ExceptionFactory, and can be overridden in
future subclasses to invoke other reconstruction methods suitable to
that particular ExceptionInfo subclass.

Signed-off-by: SZEDER Gábor <szeder@fzi.de>