Integration test to show failure when a thread throws an exception
This patch is the first in a series to support reconstruction of
exceptions thrown in a thread created by a concurrent unit test, so
it adds an integration test to show how things are working currently
with respect to exceptions thrown in child threads.  That is, such an
exception violates the NoUncaughtExceptionProperty property, resulting
in a JPFPropertyViolated exception which will cause the concurrent
unit test's failure.  At the end of the series we will amend this
test to demonstrate the new and improved behaviour.

This test is basically the same as the
ExceptionInThreadCausesPropertyViolation assumption test.  Since this
new integration test implicitly tests that assumption anyway,
there is no need for that assumption test anymore, so remove it.

Signed-off-by: SZEDER Gábor <szeder@fzi.de>