ElementInfoWrapper: search superclasses recursively for fields
Instances of JPF's ElementInfo class only contain fields and methods
actually defined in the associated class, but not those defined in its
superclass(es).  However, as we are enhancing exception reconstruction
we will soon need to access fields in superclasses (e.g. to support
the reconstruction of MultipleFailureExceptions or uncaught
exceptions).  On the other hand we have no need for a method to access
fields defined only in a particular class but not its superclasses, so
this patch enhances ElementInfoWrapper.getFieldInfo() to recursively
search all superclasses for the field with the given name until one is
found.

Signed-off-by: SZEDER Gábor <szeder@fzi.de>