ResultCollector: handle concurrency bugs and MultipleFailureExceptions
Like regular sequential unit tests, concurrent unit tests can fail
because of multiple failures, too, and one of those failures might be
a concurrency bug (e.g. the test method fails because of a concurrency
bug and an @After method fails, too).  In this case we have to report
all the failure-causing exceptions and that there was a concurrency
bug.  However, JUnit will only report multiple failures from a test if
its reporting component directly receives a MultipleFailureException
instance.  Consequently, we can't just put the whole MFE into a
ConcurrentError as cause, because the exceptions in the MFE will not
be reported.  Instead, we replace all failures stored in the MFE with
ConcurrentError exceptions, each having one of the original exceptions
as cause.

Signed-off-by: SZEDER Gábor <szeder@fzi.de>