TestWrapper: test run() method
TestWrapper.run() invokes methods of the NotifierMethods class, and
since those methods are static, we can't test that they are invoked in
various cicrumstances or that they are invoked with the proper
arguments.

Therefore, this patch introduces helper methods around those static
methods, and we can override these in tests to check invocation and
arguments.

This patch also adds two tests to check notification in case of a
succeeded and in case of a failed test.

Signed-off-by: SZEDER Gábor <szeder@fzi.de>