ExceptionInfoCollector: collect from JPF's ExceptionInfo
An exception thrown in a thread causes the violation of
NoUncaughtExceptionsProperty, which will hold the representation of
that exception in an (JPF's) ExceptionInfo instance.  This patch
extends ExceptionInfoCollector to extract the relevant information
from JPF's ExceptionInfo needed to reconstruct the violating exception
outside of JPF.

There was a catch with the stack trace of the exception: we need to be
able to gather the stack trace from two places.  The stackTrace field
contains the stack trace as a StackTraceElement[], but only if it was
initialized, either explicitly by Throwable.setStackTrace() or
implicitly by getStackTrace(), printStackTrace(), or similar.  If none
of these methods of a Throwable instance were invoked before it was
thrown without a matching catch statement, then its stackTrace field
remained uninitialized.  In such cases we must resort to some magic,
and extract the stack trace from the Throwable's snapshot field (this
magic is based on JPF's ThreadInfo.printStackTrace() method and
ThreadInfo$StackTraceElement's constructor).

The patch also adds assumption tests to check the names of those
fields of JPF's Throwable and StackTraceElement classes that we rely
upon.

Signed-off-by: SZEDER Gábor <szeder@fzi.de>