Report the exception caught by NoUncaughtExceptionsProperty
The violation of NoUncaughtExceptionsProperty was handled like any
other JPF properties.  With this patch we will treat it differently,
i.e. we will reconstruct the exception caused the violation of this
property.  This change finally makes possible to report exceptions
thrown in a non-main thread during a concurrent unit test.

Also update the related integration test, because from now on we will
report the actual uncaught exception instead of JPFPropertyViolated,
and extend it for more thorough testing.  In particular, the test will
now throw a chained exception, where the stackTrace of one of the
exceptions is initialized implicitly while the other is left
uninitialized to excercise both codepaths added in the previous commit
for gathering stack traces.

Signed-off-by: SZEDER Gábor <szeder@fzi.de>