ConcurrentRunner: forbid @Rule in concurrent test classes
JUnit release 4.7 introduced Rules.  In JUnit's current Rules
implementation @Rule fields are accessed both from the test framework
and the unit tests.  This is really bad for concurrent unit tests,
because they are run under JPF, i.e. in a JVM different from the
testing framework's JVM.  Consequently, Rules can't be used in
combination with concurrent unit tests.

So, error out when the user tries to use Rules in a concurrent test
class.

Signed-off-by: SZEDER Gábor <szeder@fzi.de>